requirements.txt
adapter-transformers==3.1.0
datasets==2.11.0
wandb==0.13.5
ruamel_yaml==0.17.21
primefac==2.0.12
cvxopt==1.3.1